# backend/api/swagger/swagger.yaml
basePath: /
definitions:
...
  auth.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    type: object
...
  secret_guest.AnswerTypesResponse:
    properties:
      answer_types:
        items:
          $ref: '#/definitions/secret_guest.AnswerTypeResponse'
        type: array
    type: object
...
  secret_guest.CreateListingRequestDTO:
    properties:
      address:
        type: string
      city:
        type: string
      code:
        type: string
      country:
        type: string
      description:
        type: string
      latitude:
        type: number
      listing_type_id:
        type: integer
      longitude:
        type: number
      main_picture:
        type: string
      title:
        type: string
    required:
    - address
    - city
    - code
    - country
    - description
    - latitude
    - listing_type_id
    - longitude
    - title
    type: object
...
  secret_guest.ErrorResponse:
    properties:
      message:
        example: An unexpected error occurred.
        type: string
    type: object
...
  secret_guest.ListingResponseDTO:
    properties:
      address:
        type: string
      city:
        type: string
      code:
        type: string
      country:
        type: string
      description:
        type: string
      id:
        type: string
      latitude:
        type: number
      listing_type:
        $ref: '#/definitions/secret_guest.ListingTypeResponse'
      longitude:
        type: number
      main_picture:
        type: string
      title:
        type: string
    type: object
...
  secret_guest.OTAReservationRequestDTO:
    properties:
      received_at:
        type: string
      reservation:
        $ref: '#/definitions/secret_guest.OTAReservationDTO'
      source:
        type: string
    required:
    - received_at
    - reservation
    - source
    type: object
...
info:
  contact: {}
  description: This is the API for the Secret Guest application.
  title: Secret Guest API
  version: "1.0"
paths:
  /admin/listings:
    post:
      consumes:
      - application/json
      description: Creates new listing object. Available for admin only.
      parameters:
      - description: Listing Payload
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/secret_guest.CreateListingRequestDTO'
      - description: Bearer Access Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/secret_guest.ListingResponseDTO'
        "400":
          description: Invalid payload
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
        "409":
          description: Listing cannot be created (e.g., duplicate listing or conflict)
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create new Listing (Admin)
      tags:
      - Listings (Admin)
  /admin/sg_reservations:
    post:
      consumes:
      - application/json
      description: Handles OTA Reservations Events. Available for admin only.
      parameters:
      - description: OTA Reservation Payload
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/secret_guest.OTAReservationRequestDTO'
      - description: Bearer Access Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Invalid payload
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Handles OTA Reservations Events (Admin)
      tags:
      - Reservations (Admin)
  /answer_types:
    get:
      description: Returns a list of all Answer Types with optional filtering.
      parameters:
      - collectionFormat: multi
        description: Filter by one or more Answer Type IDs
        in: query
        items:
          type: integer
        name: id
        type: array
      - collectionFormat: multi
        description: Filter by one or more Answer Type slugs
        in: query
        items:
          type: string
        name: slug
        type: array
      - description: Bearer Access Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/secret_guest.AnswerTypesResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Answer Types (Staff)
      tags:
      - Answer Types (Staff)
    post:
      consumes:
      - application/json
      description: Creates a new Answer Type.
      parameters:
      - description: Answer Type Payload
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/secret_guest.CreateAnswerTypeRequestDTO'
      - description: Bearer Access Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/secret_guest.AnswerTypeResponse'
        "400":
          description: Invalid payload
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
        "409":
          description: Duplicate slug
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Answer Type (Staff)
      tags:
      - Answer Types (Staff)
...
  /assignments:
    get:
      description: Returns a paginated list of all assignments with optional filters.
        Available for staff only.
      parameters:
      - default: 1
        description: Page number for pagination
        in: query
        name: page
        type: integer
      - default: 50
        description: Number of items per page
        in: query
        name: limit
        type: integer
      - description: Filter by reporter (user) ID
        format: uuid
        in: query
        name: reporter_id
        type: string
      - collectionFormat: multi
        description: Filter by one or more status IDs
        in: query
        items:
          type: integer
        name: status_id
        type: array
      - description: Bearer Access Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/secret_guest.AssignmentsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get All Assignments (Staff)
      tags:
      - Assignments (Staff)
  /assignments/{id}:
    get:
      description: Returns details of any assignment. Available for staff only.
      parameters:
      - description: Assignment ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Bearer Access Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/secret_guest.AssignmentResponseDTO'
        "400":
          description: Invalid assignment ID format
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
        "404":
          description: Assignment not found
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Assignment By ID (Staff)
      tags:
      - Assignments (Staff)
  /assignments/{id}/take:
    patch:
      description: Allows a user to take a free assignment, assigning it to themselves.
        The assignment status becomes 'offered' to this specific user.
      parameters:
      - description: Assignment ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Bearer Access Token
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid assignment ID format
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
        "404":
          description: Assignment not found or not available
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
        "409":
          description: Assignment cannot be taken (e.g., already taken, or user has
            other active offers)
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Take a Free Assignment
      tags:
      - Assignments (User)
  /assignments/my:
    get:
      description: Returns a paginated list of assignments for the current user.
      parameters:
      - default: 1
        description: Page number for pagination
        in: query
        name: page
        type: integer
      - default: 20
        description: Number of items per page
        in: query
        name: limit
        type: integer
      - description: Bearer Access Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/secret_guest.AssignmentsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get My Assignments
      tags:
      - Assignments (User)
  /assignments/my/{id}:
    get:
      description: Returns details of a specific assignment belonging to the current
        user.
      parameters:
      - description: Assignment ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Bearer Access Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/secret_guest.AssignmentResponseDTO'
        "400":
          description: Invalid assignment ID format
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
        "404":
          description: Assignment not found or does not belong to user
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get My Assignment By ID
      tags:
      - Assignments (User)
  /assignments/my/{id}/accept:
    patch:
      description: Accept an assignment offer. This will start an asynchronous process
        to generate a report.
      parameters:
      - description: Assignment ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Bearer Access Token
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "202":
          description: Accepted
        "400":
          description: Invalid assignment ID format
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
        "404":
          description: Assignment not found or does not belong to user
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
        "409":
          description: Assignment cannot be accepted (e.g., wrong status)
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Accept My Assignment
      tags:
      - Assignments (User)
  /assignments/my/{id}/decline:
    patch:
      description: Decline an assignment offer.
      parameters:
      - description: Assignment ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Bearer Access Token
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid assignment ID format
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
        "404":
          description: Assignment not found or does not belong to user
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
        "409":
          description: Assignment cannot be declined (e.g., wrong status)
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Decline My Assignment
      tags:
      - Assignments (User)
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Refresh a JWT token using a refresh token
      parameters:
      - description: Refresh Token
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/auth.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.RefreshTokenResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: Invalid or expired refresh token
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Refresh an existing JWT token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a user with username, password, and email
      parameters:
      - description: User Registration Info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 'message: User registered successfully'
          schema:
            $ref: '#/definitions/auth.RegisterUserResponse'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "409":
          description: User with this username or email already exists
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
  /auth/token:
    post:
      consumes:
      - application/json
      description: Generates a new JWT access token for authentication.
      parameters:
      - description: User Credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/auth.GenerateTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.GenerateTokenResponse'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: Invalid username or password
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Generate a new access token
      tags:
      - auth
  /auth/validate:
    post:
      description: Validates the provided JWT access token.
      parameters:
      - description: Bearer Access Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.ValidateTokenResponse'
        "401":
          description: Unauthorized or invalid token
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Validate an access token
      tags:
      - auth
  /checklist_items:
    get:
      description: Returns a list of all Checklist Items.
      parameters:
      - collectionFormat: multi
        description: Filter by one or more Checklist Item IDs
        in: query
        items:
          type: integer
        name: id
        type: array
      - collectionFormat: multi
        description: Filter by one or more Checklist Item slugs
        in: query
        items:
          type: string
        name: slug
        type: array
      - collectionFormat: multi
        description: Filter by one or more Listing Type IDs
        in: query
        items:
          type: integer
        name: listing_type_id
        type: array
      - collectionFormat: multi
        description: Filter by one or more Listing Type slugs
        in: query
        items:
          type: string
        name: listing_type_slug
        type: array
      - description: Filter by active status
        in: query
        name: is_active
        type: boolean
      - description: Bearer Access Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/secret_guest.ChecklistItemsResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Checklist Items (Staff)
      tags:
      - Checklist Items (Staff)
    post:
      consumes:
      - application/json
      description: Creates a new Checklist Item.
      parameters:
      - description: Checklist Item Payload
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/secret_guest.CreateChecklistItemRequestDTO'
      - description: Bearer Access Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/secret_guest.ChecklistItemResponse'
        "400":
          description: Invalid payload or non-existent foreign key
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
        "409":
          description: Duplicate slug for this listing type
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Checklist Item (Staff)
      tags:
      - Checklist Items (Staff)
...
  /listing_types:
    get:
      description: Returns a list of all Listing Types.
      parameters:
      - collectionFormat: multi
        description: Filter by one or more Listing Type IDs
        in: query
        items:
          type: integer
        name: id
        type: array
      - collectionFormat: multi
        description: Filter by one or more Listing Type slugs
        in: query
        items:
          type: string
        name: slug
        type: array
      - description: Bearer Access Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/secret_guest.ListingTypesResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Listing Types (Staff)
      tags:
      - Listing Types (Staff)
    post:
      consumes:
      - application/json
      description: Creates a new Listing Type.
      parameters:
      - description: Listing Type Payload
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/secret_guest.CreateListingTypeRequestDTO'
      - description: Bearer Access Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/secret_guest.ListingTypeResponse'
        "400":
          description: Invalid payload
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
        "409":
          description: Duplicate slug
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/secret_guest.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Listing Type (Staff)
      tags:
      - Listing Types (Staff)
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and a JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"